JBIDE-14994 - Page differences between page displayed via LiveReload and original page sent from server
Proxy server was not forwarding the full URI, it used to remove the QueryParams, which would cause the seam app on AS 7.1 to return 404
Added a test to verify that, using a dedicated Servlet in the Preview Server. This servlet will return a 400 error if the incoming request
did not contain query params.