JBIDE-15613 - Error: Failed to send reload command to browser
Problem was that the ServerListener would send too many notifications
when a module would be published. Once the browser received the first
notification, it would refresh the page, which would close the current
websocket. Further messages for the same module publication would be sent
by the plugin, but one of them would fail to be delivered because the connection
would be closed.

Modified the JUnit test and HTTP Preview Server to reproduce the multiple events
and assert that only one notification is sent.
Adjusted some log messages as well.