JBIDE-14999 - LiveReload: add option for refresh delay
Added a spinner in the Server Configuration Editor to set
the number of seconds that the livereload server should wait after
the application server published the resources that changed.

No need to restart the server to take the new delay into account (just
save the change).
The delay is seen as a job in the Progress View, so the user can cancel
the job is the delay was too long.

Added JUnit test.