JBIDE-21117 - Livereload tests fail after migration to jetty 9.3.5
Fixing broken tests after upgrading to Jetty 9.3.5
Servlet Filter that injects the livereload.js script in
the HTML response would not work and return an empty
response.
Replaced with an AsyncMiddleManServlet that provides a
custom ContentTransformer when script injection is enabled
and should happen.
Some refactoring on the WorkspaceServlet that serves content
on a specific path and that is proxied by the servlet aforementionned
to remove the faulty servlet filter in all use-cases.